requirements.txt

python-dotenv
pandas
openpyxl
python-dotenv==1.0.0
python-telegram-bot==21.7
supabase==2.3.0
